backend/requirements.txt
Flask==2.3.2
flasgger==0.9.7.1
flask_sqlalchemy==3.1.1
flask_restful   
psycopg2-binary